Added hints for how to add newly published models
When Redox publishes a brand new set of models, there won't be name
translations that exist for them, which means the lib generator won't
know how to name files and classes in the generated code.

For someone going through the CI log to figure out why a build failed,
these changes will identify this problem and give suggestions for how to
remedy the situation.

Also, this commit includes name translations for a newly published model
(Enrichment).